chore: rename UI
