+++ apps/embed/package.json
@@ -28,15 +28,15 @@
     "@types/react": "18.2.13",
     "@types/react-dom": "18.2.6",
     "autoprefixer": "^10.4.14",
-    "@lenstube/config": "*",
-    "@lenstube/lens": "*",
     "postcss": "^8.4.24",
     "prettier": "^2.8.8",
     "tailwindcss": "^3.3.2",
     "typescript": "5.1.3",
+    "@lenstube/config": "*",
+    "@lenstube/lens": "*",
     "@lenstube/browser": "*",
     "@lenstube/generic": "*",
     "@lenstube/constants": "*",
-    "web-ui": "*"
+    "@lenstube/ui": "*"
   }
 }
